MAKE ALL REST API FUNCTIONS WORK WITHOUT AUTHORIZATION / AUTHENTICATION
All required API functions now work - GET, POST, PUT & DELETE - albeit without the client or the user
being required to authenticate or authorize.
This will be addressed next.

Had to remove the .env composer module - this seemed to cause a SQL error
randomly because the password had not been supplied.
Replaced .env with common define CONSTANTS boostrapped from a env.php
file and this seemed to resolve the isse.