ADD TOKEN HANDLING CLASS FOR REFRESH AND ACCESS TOKENS AND ADD CHECKS IN AUTHORIZED ROUTES
Created a TokenHandler Class with STATIC methods to be called from the AUTH module

Added Functionality to create Jason Web tokens for a short lived Access Token (30 mins)
and a long lived Refresh Token (30 days)

Set up database for blacklisted Refresh tokens so that if token is compromised
it can be added to the blacklist and grant of Access Token denied - so giving
maximum 'risk' window of 30 minutes from the grant of the last Access token
to blocking access.

Added a Refresh route to request new Access token and tidied up the front end
to check for auth routes for login / register / refresh.

Decided to use http only cookies to hold the JWT since these are sent automatically
with each fetch request (if credentials are enabled) and they are not accessibly by
Javascript - so reducing risk of XSS attack.

Also set samesite = strict to reduce risk of CSRF type attack but if was being very cautious
would add a CSRF token generator path to be included in form submisions from the App.
However this is just a low security demo API - so will not implement.

Also added a standard Cookie called RefreshTimer that can be accessed by Javascript
and has the same lifespan as the AccessToken to allow the Vue app to check if AccessToken
has expired before making fetch requests.