ADD A LOGOUT ROUTE AND CLEAN UP CODE
A logout route was added to destroy refresh and access tokens.

Code was cleaned up to make it more readable.

API is now working as it should.

So long as the API is on the same uri / server as the request App
then there are no CORS issues. If not then coudl eb some issues
with browser accepting http cookies and may need to rework to use
header authorization and bearer tokens instead.